some fixes

+++ src/ags/audio/ags_recall_id.h
@@ -30,14 +30,14 @@
 #define AGS_IS_RECALL_ID(obj)             (G_TYPE_CHECK_INSTANCE_TYPE ((obj), AGS_TYPE_RECALL_ID))
 #define AGS_IS_RECALL_ID_CLASS(class)     (G_TYPE_CHECK_CLASS_TYPE ((class), AGS_TYPE_RECALL_ID))
 #define AGS_RECALL_ID_GET_CLASS(obj)      (G_TYPE_INSTANCE_GET_CLASS ((obj), AGS_TYPE_RECALL_ID, AgsRecallIDClass))
-#define AGS_POINTER_TO_GROUP_ID(ptr)      ((AgsGroupId)(gulong)(ptr))
+#define AGS_POINTER_TO_GROUP_ID(ptr)      ((AgsGroupId)(guint)(ptr))
 #define AGS_GROUP_ID_TO_POINTER(id)       ((gpointer)(AgsGroupId)(id))
 
 #define AGS_GROUP_ID_NOT_EXISTIBLE 0
 
 typedef struct _AgsRecallID AgsRecallID;
 typedef struct _AgsRecallIDClass AgsRecallIDClass;
-typedef gulong AgsGroupId;
+typedef guint AgsGroupId;
 
 typedef enum{
   AGS_RECALL_ID_RUN_PRE_SYNC_ASYNC_DONE     = 1,
